Reworked dataset list
